rest/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-login
python-dotenv
psycopg2-binary